Implementation of the following functions: splitBase-returns the base
splitNum-return the number at that base

int2Number- returns the converted number as the given base. use "BuildNum" to return it as a string <number>b<base>

is number-checks if the string meets the conditions

number2Int-if the string is good returns it value as 10 base

crating "equals" and using it in -"splitNum","splitBase" and "isNumber"

BuildNum-bulding a string that represent as needed the number and it base

maxIndex- check between the 2 numbers, the add & multification, and between the max of 2 of them.

add-Calculates the sum between them and returns as <number>b<base>

multi-Calculates the multification between them and returns as <number>b<base>